chore: split platform related into seprate files
